draft: A list of people, books and other things.
Really a very early draft, I haven't yet though about the implications
of those things, but it promises to be quite interesting.

In the long term, I'd rather have this a page, not an article, but this
little thing doesn't support them yet. And there are a few things I'd
like to try out related to this kind of lists/linkage.