Sync script tags mess up things.
In fact, in this particular post they have the positive effect that the
stats are rendered directly after the example, but also the negative
effect that they don't include the complete post, just the nodes in up
to the script tag, which is ... interesting.

When I say, "the stats are incorrect" I am aware that they would be
incorrect anyway, because of MANY THINGS. But still, even if the
"algorithm" were perfect, the stats would still be wrong because they
don't include all the data.

I don't think I should fix the problem for this little script, because
I think it makes a much better example and maybe provokes someone to dig
deeper and discover something new about the pitfalls of HTML rendering.

A "solution" that provides both the correct (see above) stats and
demonstrates synchronous execution of script tags could be the ability
to execute that snippet again and probably overwrite the message. But
not now.

Whew, that's almost a post by itself. I should gather stats about the
length of my commit messages, too. :)

See you.